LayoutTests/js/array-iterators-expected.txt

This test checks the behavior of the iterator methods on Array objects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS key is 0
PASS key is 1
PASS key is 2
PASS key is 3
PASS key is 4
PASS key is 5
PASS testArray.length is 6
PASS value is 1
PASS value is 2
PASS value is 3
PASS value is 4
PASS value is 5
PASS value is 6
PASS testArray.length is 6
PASS value is testArray[key]
PASS key is 0
PASS value is 1
PASS value is testArray[key]
PASS key is 1
PASS value is 2
PASS value is testArray[key]
PASS key is 2
PASS value is 3
PASS value is testArray[key]
PASS key is 3
PASS value is 4
PASS value is testArray[key]
PASS key is 4
PASS value is 5
PASS value is testArray[key]
PASS key is 5
PASS value is 6
PASS testArray.length is 6
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS testArray.length is 9
PASS successfullyParsed is true

TEST COMPLETE
